Tarefa2_2022li1g079

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa :: Mapa -> Int -> Mapa

A função estendeMapa é responsável por gerar novas linhas válidas para o mapa, assim como obstaculos para as novas linhas. Para o fazer utiliza as funções auxiliares geradorT e geradorObs

geradorT :: Mapa -> Int -> Terreno

geradorObs :: Terreno -> [Obstaculo] -> Int -> Int -> [Obstaculo]

proximosTerrenosValidos :: Mapa -> [Terreno]

A função proximosTerrenosValido recebe um mapa e dá uma lista de todos os terrenos seguintes possiveis

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

A função proximosObstaculosValidos é responsável por gerar a lista de possoveis seguintes obstaculos, respeitando as restrições. Para isso utiliza uma função auxiliar proximosObstaculosValidosaux

proximosObstaculosValidosaux :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

A função proximosObstaculosValidosaux é a função auxiliar da função proximosObstaculosValidos, e é responsável por restringir a nova lista de obstaculos tendo em conta as restrições dadas