Tarefa3_2022li1g079

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo :: Jogo -> Jogada -> Jogo

A função animaJogo é responsável por movimentar o jogador (segundo a jogada) e os obstáculos (segundo as velocidades)

player :: Jogador -> Mapa -> Jogada -> Jogador

a função player determina as próximas coordenadas do jogador, consoante a sua posição no mapa e a existência de obstaculos, na primeira linha é verificado que se o jogador se encontrar em cima de um tronco e não efetuar nenhuma jogada (Parado), movimenta-se consoante a velocidade do tronco.

Obs: na ultima linha o "otherwise" representa o movimento Esquerda

posT :: Mapa -> Coordenadas -> [Coordenadas]

a função posT determina as coordenadas de todos os troncos num dado mapa, para o fazer utiliza a função auxiliar posTaux

posTaux :: (Terreno, [Obstaculo]) -> Coordenadas -> [Coordenadas]

a função posTaux é uma função auxiliar que determina as coordenadas de todos os troncos numa linha do mapa

posA :: Mapa -> Coordenadas -> [Coordenadas]

a função posA determina as coordenadas de todas as árvores num dado mapa, para o fazer utiliza a função auxiliar posAaux

posAaux :: (Terreno, [Obstaculo]) -> Coordenadas -> [Coordenadas]

a função posAaux é uma função auxiliar que determina as coordenadas de árvores numa linha do mapa

posMoveT :: Coordenadas -> Mapa -> Coordenadas

a função posMoveT é a função responsável por obter as proximas coordenadas de um tronco, para mais tarde ser usado no caso do jogador estar em cima de um tronco, se movimentar com ele

animaObs :: Mapa -> Mapa

A função posMoveVelocidadeAux é a função auxiliar da função posMoveVelocidade, responsável por obter as coordenadas do tronco consoante a velocidade Obs: Só utiliza um Int pois os troncos movimentam-se na horizontal, logo só vai mudar o y

a função animaObs é responsável por movimentar todos os obstáculos presentes num dado mapa, utiliza para isso uma função auxiliar animaObsAux

animaObsAux :: [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]

a função animaObsAux é uma função auxiliar da função animaObs e é responsável por movimentar os obstaculos de uma lista de (Terreno,[Obstaculo]) tendo em conta a velocidade do Terreno, e fazendo os obstaculos que desaparecem de um lado do mapa aparecerem do lado oposto